+++ demosplan/DemosPlanCoreBundle/Logic/Document/SingleDocumentService.php
@@ -14,6 +14,8 @@ use DemosEurope\DemosplanAddon\Contracts\Entities\SingleDocumentInterface;
 use DemosEurope\DemosplanAddon\Contracts\Services\SingleDocumentServiceInterface;
 use demosplan\DemosPlanCoreBundle\Entity\Document\SingleDocument;
 use demosplan\DemosPlanCoreBundle\Entity\Document\SingleDocumentVersion;
+use demosplan\DemosPlanCoreBundle\Entity\Report\ReportEntry;
+use demosplan\DemosPlanCoreBundle\Event\CreateReportEntryEvent;
 use demosplan\DemosPlanCoreBundle\Logic\CoreService;
 use demosplan\DemosPlanCoreBundle\Logic\DateHelper;
 use demosplan\DemosPlanCoreBundle\Logic\EntityHelper;
@@ -26,6 +28,8 @@ use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use Exception;
 use ReflectionException;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
+
 
 class SingleDocumentService extends CoreService implements SingleDocumentServiceInterface
 {
@@ -42,6 +46,7 @@ class SingleDocumentService extends CoreService implements SingleDocumentService
         private readonly SingleDocumentVersionRepository $singleDocumentVersionRepository,
         private readonly SingleDocumentReportEntryFactory $reportEntryFactory,
         private readonly ReportService $reportService,
+        private readonly EventDispatcherInterface $eventDispatcher,
     ) {
         $this->fileService = $fileService;
     }
@@ -237,8 +242,8 @@ class SingleDocumentService extends CoreService implements SingleDocumentService
     {
         $singleDocument = $this->singleDocumentRepository->add($data);
 
-        $report = $this->reportEntryFactory->createSingleDocumentCreateEntry($singleDocument);
-        $this->reportService->persistAndFlushReportEntries($report);
+        $reportEntryEvent = new CreateReportEntryEvent($singleDocument, ReportEntry::CATEGORY_ADD);
+        $this->eventDispatcher->dispatch($reportEntryEvent);
 
         $singleDocument = $this->entityHelper->toArray($singleDocument);
         $singleDocument = $this->convertDateTime($singleDocument);
@@ -264,9 +269,11 @@ class SingleDocumentService extends CoreService implements SingleDocumentService
             foreach ($idents as $documentId) {
                 try {
                     $documentToDelete = $this->getSingleDocument($documentId, false);
-                    $report = $this->reportEntryFactory->createSingleDocumentDeleteEntry($documentToDelete);
+
+                    $reportEntryEvent = new CreateReportEntryEvent($documentToDelete, ReportEntry::CATEGORY_DELETE);
+                    $this->eventDispatcher->dispatch($reportEntryEvent);
+
                     $this->singleDocumentRepository->delete($documentId);
-                    $this->reportService->persistAndFlushReportEntries($report);
                 } catch (Exception $e) {
                     $this->logger->error('Fehler beim Löschen eines SingleDocuments: ', [$e]);
                     $success = false;
@@ -291,8 +298,8 @@ class SingleDocumentService extends CoreService implements SingleDocumentService
     {
         $updatedDocument = $this->singleDocumentRepository->update($data['ident'], $data);
 
-        $report = $this->reportEntryFactory->createSingleDocumentUpdateEntry($updatedDocument);
-        $this->reportService->persistAndFlushReportEntries($report);
+        $reportEntryEvent = new CreateReportEntryEvent($updatedDocument, ReportEntry::CATEGORY_UPDATE);
+        $this->eventDispatcher->dispatch($reportEntryEvent);
 
         $updatedDocument = $this->entityHelper->toArray($updatedDocument);
         $updatedDocument = $this->convertDateTime($updatedDocument);