+++ demosplan/DemosPlanCoreBundle/Logic/Report/PlanDrawReportEntryFactory.php
@@ -18,6 +18,7 @@ use demosplan\DemosPlanCoreBundle\Entity\Report\ReportEntry;
 use demosplan\DemosPlanCoreBundle\Entity\User\AnonymousUser;
 use demosplan\DemosPlanCoreBundle\Entity\User\User;
 use demosplan\DemosPlanCoreBundle\Exception\UserNotFoundException;
+use function PHPUnit\Framework\assertNotNull;
 
 class PlanDrawReportEntryFactory extends AbstractReportEntryFactory
 {
@@ -46,6 +47,7 @@ class PlanDrawReportEntryFactory extends AbstractReportEntryFactory
         string $oldPlanDrawPDF,
         string $newPlanDrawPDF,
     ): ReportEntry {
+        $messageData = null;
 
         if (0 !== strcmp($oldPlanPDF, $newPlanPDF)) {
             $messageData['planDrawFile']['old'] = $oldPlanPDF;
@@ -56,6 +58,8 @@ class PlanDrawReportEntryFactory extends AbstractReportEntryFactory
             $messageData['planDrawingExplanation']['old'] = $oldPlanDrawPDF;
             $messageData['planDrawingExplanation']['new'] = $newPlanDrawPDF;
         }
+
+        assertNotNull($messageData);
 
         $reportEntry = $this->createPlanDrawReportEntry($procedureId, $messageData);
         $reportEntry->setCategory(ReportEntry::CATEGORY_CHANGE);