+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+- Create report entries on create, update, deletion of an element, paragraph , singleDocument, mapDrawing or mapDrawing-explanation
 - Add the possibility to export synopsis without personal data.
 
 ## v2.26.0 (2025-02-25)